Rendering with filter + opacity

+++ Plugin/Parameters.h
@@ -12,5 +12,5 @@ public:
 	float defaultSpeed;
 	bool destroyOutsideScreen;
 	bool useTexture;
-	int bulletSize;
+	float bulletSize;
 };
